Hangwei_BackEnd/hangwei/requirements.txt
fuzzychinese==0.1.5
setuptools==20.7.0
requests==2.9.1
redis==3.2.1
jieba==0.39
pandas==0.24.2
baidu_aip==2.2.13.0
numpy==1.16.3
gensim==3.7.3
Django==2.2.1
pypinyin==0.35.3
ChineseTone==0.1.4
Pillow==6.0.0
seaborn==0.9.0
scikit_learn==0.21.1
